requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pyautogui==0.9.54
pillow==10.0.0
mss==9.0.1
pytesseract==0.3.10 